Share one direct upload size check across all entry points
createFileUpload and the application file batch each duplicated the
maxDirectUploadFileSize check that FileUploadTargetService now enforces.
Extract it to assertDirectUploadSizeOrThrow: createFileUpload still calls
it before creating the pending row, and the batch validation drops its
copy since the target service rejects oversized files per file.
