+++ packages/twenty-server/src/engine/core-modules/file/file-upload/services/file-upload-target.service.ts
@@ -1,13 +1,11 @@
 import { Injectable } from '@nestjs/common';
 
 import { msg } from '@lingui/core/macro';
-import bytes from 'bytes';
 
 import { ApiPath, FileFolder } from 'twenty-shared/types';
 import { isDefined } from 'twenty-shared/utils';
 import { v4 } from 'uuid';
 
-import { settings } from 'src/engine/constants/settings';
 import { FileStorageService } from 'src/engine/core-modules/file-storage/services/file-storage.service';
 import { validateFilePath } from 'src/engine/core-modules/file-storage/utils/validate-file-path.util';
 import { FileUploadTargetDTO } from 'src/engine/core-modules/file/file-upload/dtos/file-upload-target.dto';
@@ -16,6 +14,7 @@ import {
   FileUploadExceptionCode,
 } from 'src/engine/core-modules/file/file-upload/file-upload.exception';
 import { type BatchFileResult } from 'src/engine/core-modules/file/file-upload/types/batch-file-result.type';
+import { assertDirectUploadSizeOrThrow } from 'src/engine/core-modules/file/file-upload/utils/assert-direct-upload-size-or-throw.util';
 import { buildPendingUploadResourcePath } from 'src/engine/core-modules/file/file-upload/utils/build-pending-upload-resource-path.util';
 import { toBatchErrorMessage } from 'src/engine/core-modules/file/file-upload/utils/to-batch-error-message.util';
 import { FileSettings } from 'src/engine/core-modules/file/types/file-settings.types';
@@ -61,7 +60,7 @@ export class FileUploadTargetService {
     contentType: string;
     size: number;
   }): Promise<FileUploadTargetDTO> {
-    this.assertUploadSizeAllowedOrThrow(size);
+    assertDirectUploadSizeOrThrow(size);
 
     const pendingResourcePath = buildPendingUploadResourcePath({
       fileId,
@@ -139,7 +138,7 @@ export class FileUploadTargetService {
         try {
           // Checked before the pending row is created: buildUploadTarget would
           // reject it too, but only after leaving a row for the cron to reap.
-          this.assertUploadSizeAllowedOrThrow(request.size);
+          assertDirectUploadSizeOrThrow(request.size);
 
           const pendingFile = await this.fileStorageService.createPendingFile({
             fileFolder: request.fileFolder,
@@ -171,19 +170,5 @@ export class FileUploadTargetService {
         }
       }),
     );
-  }
-
-  private assertUploadSizeAllowedOrThrow(size: number): void {
-    const maxFileSize = bytes(settings.storage.maxDirectUploadFileSize) ?? 0;
-
-    if (!Number.isInteger(size) || size <= 0 || size > maxFileSize) {
-      throw new FileUploadException(
-        `Invalid file size ${size} (max ${maxFileSize} bytes)`,
-        FileUploadExceptionCode.FILE_TOO_LARGE,
-        {
-          userFriendlyMessage: msg`The file is empty or exceeds the maximum allowed size.`,
-        },
-      );
-    }
   }
 }
